requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
feedparser==6.0.11
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
pymongo==4.14.1
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
sgmllib3k==1.0.0
tweepy==4.16.0
urllib3==2.5.0
vaderSentiment==3.3.2
Werkzeug==3.1.3
